+++ grammar.js
@@ -1,13 +1,6 @@
-"use strict";
-// opencypher tests
-// https://github.com/opencypher/openCypher/blob/c816756d50df9cde73cae573ef871f2d7e76c70a/tools/grammar/src/test/resources/cypher.txt
-const OCT_INT = /0o(_?[0-7])+/;
-const HEX_INT = /0x(_?[0-9a-f])+/i;
-const DEC_INT = /[0-9](_?[0-9])*/;
-const DECIMAL = token.immediate(seq(optional(DEC_INT), '.', DEC_INT));
-const DECIMAL_SCIENTIFIC = token.immediate(seq(choice(DEC_INT, DECIMAL), choice('e', 'E'), optional(choice('-', '+')), DEC_INT));
-const DECIMAL_LITERAL = token.immediate(seq(choice(DECIMAL, DECIMAL_SCIENTIFIC), optional(choice('f', 'F', 'd', 'D'))));
-module.exports = grammar({
+import { OCT_INT, HEX_INT, DEC_INT, DECIMAL_LITERAL } from './grammar/num.js';
+import { REGULAR_IDENTIFIER, EXTENDED_IDENTIFIER } from './grammar/ident.js';
+export default grammar({
     name: 'cypher',
     extras: ($) => [$.comment, $._whitespace_char],
     conflicts: () => [],
@@ -16,6 +9,11 @@ module.exports = grammar({
         $.variable_in_parens,
         $.double_quote_string,
         $.single_quote_string,
+        $.delimited_identifier, // TODO: Verify which nodes should be inlined.
+        $.non_reserved_word, // TODO: Verify which nodes should be inlined.
+        // $.identifier,        // TODO: Verify which nodes should be inlined.
+        $.symbolic_name,
+        $.schema_name,
     ],
     rules: {
         cypher: ($) => seq($.statement, optional(';')),
@@ -133,13 +131,15 @@ module.exports = grammar({
         list_literal: ($) => seq('[', optional(commaSeparated($.expression)), ']'),
         map_literal: ($) => seq('{', optional(commaSeparated(seq($.property_key_name, ':', $.expression))), '}'),
         property_key_name: ($) => $.schema_name,
+        /* --- Identifiers --- */
+        parameter_name: ($) => seq('$', choice(EXTENDED_IDENTIFIER, $.delimited_identifier)),
+        identifier: ($) => choice(REGULAR_IDENTIFIER, $.delimited_identifier, $.non_reserved_word),
+        delimited_identifier: ($) => seq('`', repeat(choice(/[^`\\]+/, $.escaped_char, alias('``', $.escaped_char))), '`'),
+        non_reserved_word: ($) => prec(-1, choice(// TODO: verify precedence works ok.
+        word('allshortestpaths'), word('all'), word('and'), word('any'), word('as'), word('asc'), word('ascending'), word('by'), word('call'), word('case'), word('contains'), word('count'), word('create'), word('delete'), word('desc'), word('descending'), word('detach'), word('distinct'), word('else'), word('end'), word('ends'), word('exists'), word('false'), word('group'), word('groups'), word('in'), word('inf'), word('infinity'), word('is'), word('limit'), word('match'), word('merge'), word('nan'), word('none'), word('not'), word('null'), word('offset'), word('on'), word('optional'), word('or'), word('order'), word('path'), word('paths'), word('reduce'), word('remove'), word('return'), word('set'), word('shortest'), word('shortestpath'), word('single'), word('skip'), word('starts'), word('then'), word('trim'), word('true'), word('union'), word('unwind'), word('when'), word('where'), word('with'), word('xor'), word('yield'))),
         /* --- END --- */
-        schema_name: ($) => choice($.symbolic_name, $.reserved_word), // ??
-        reserved_word: () => choice(// ??
-        word('all'), word('asc'), word('ascending'), word('by'), word('create'), word('delete'), word('desc'), word('descending'), word('detach'), word('exists'), word('limit'), word('match'), word('merge'), word('on'), word('optional'), word('order'), word('remove'), word('return'), word('set'), word('skip'), word('where'), word('with'), word('union'), word('unwind'), word('and'), word('as'), word('contains'), word('distinct'), word('ends'), word('in'), word('is'), word('not'), word('or'), word('starts'), word('xor'), word('false'), word('true'), word('null'), word('constraint'), word('unique'), word('case'), word('when'), word('then'), word('else'), word('end'), word('mandatory'), word('scalar'), word('of'), word('add'), word('drop')),
-        symbolic_name: ($) => prec.left(choice($.unescaped_symbolic_name, $.escaped_symbolic_name, word('count'), word('filter'), word('extract'), word('any'), word('none'), word('single'))),
-        unescaped_symbolic_name: ($) => (/(\p{ID_Start}|\p{Pc})(\p{ID_Continue}|\p{Sc})*/u),
-        escaped_symbolic_name: () => repeat1(/`[^`]*`/),
+        schema_name: ($) => $.identifier, // ??
+        symbolic_name: ($) => $.identifier,
         comment: ($) => choice(seq('/*', repeat(choice(/[^\*]/, /\*[^\/]/)), '*/'), seq('//', /.*/, '\n')),
         left_arrow_head: () => choice('<', '\u{27e8}', '\u{3008}', '\u{fe64}', '\u{ff1c}'),
         right_arrow_head: () => choice('>', '\u{27e9}', '\u{3009}', '\u{fe65}', '\u{ff1e}'),