+++ grammar/ident.js
@@ -1,0 +1,2 @@
+export const REGULAR_IDENTIFIER = /\p{XID_Start}\p{XID_Continue}*/u;
+export const EXTENDED_IDENTIFIER = /\p{XID_Continue}+/u;